requirement.txt
streamlit
opencv-python-headless
pillow
scikit-learn
joblib